api/requirements.txt
fastapi
uvicorn
gunicorn
requests
cassandra-driver
dateparser
pandas
requests_html
selenium
bcrypt